move all the necessary logic into ALT
