+++ src/lib/alt/index.ts
@@ -58,7 +58,7 @@ export class Alt {
     };
   }
 
-  async createAndFillAlt<
+  private async createAndFillAlt<
     T extends { accounts: Array<web3.PublicKey>; altTable?: web3.PublicKey },
   >(instance: T, ty: string) {
     const createTable = await this.createTable(instance.accounts);
@@ -72,6 +72,40 @@ export class Alt {
       this.logger,
       this.baseApp.keypair,
     );
+  }
+
+  async createAndFillAltIfNecessary<
+    T extends { accounts: Array<web3.PublicKey>; altTable?: web3.PublicKey },
+  >(instance: T, ty: string) {
+    if (instance.altTable === undefined) {
+      await this.createAndFillAlt(instance, ty);
+    } else {
+      const lookupTableAccount = (
+        await this.baseApp.anchorProvider.connection.getAddressLookupTable(
+          new web3.PublicKey(instance.altTable!),
+        )
+      ).value;
+      let expectedAccounts = [...instance.accounts];
+      this.logger.info(`${ty} ALT Table Defined -- ${instance.altTable!}`);
+
+      for (let i = 1; i <= lookupTableAccount.state.addresses.length; i += 1) {
+        const lookupAddress = lookupTableAccount.state.addresses[i - 1];
+        this.logger.info(
+          `ALT Account ${i}/${lookupTableAccount.state.addresses.length} -- ${lookupAddress}`,
+        );
+        expectedAccounts = expectedAccounts.filter(
+          (account) => account.toBase58() !== lookupAddress.toBase58(),
+        );
+      }
+      if (expectedAccounts.length !== 0) {
+        for (const remainingAccount of expectedAccounts) {
+          this.logger.warn(`${ty} ALT missing -- ${remainingAccount}`);
+        }
+        this.logger.info(`${ty} Creating a new ALT`);
+        await this.createAndFillAlt(instance, ty);
+        this.logger.info(`${ty} Using new ALT -- ${instance.altTable!}`);
+      }
+    }
   }
 }
